requirements.txt
openai==1.3.0
httpx==0.24.1
pillow==10.2.0
tqdm==4.67.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
azure-ai-vision-imageanalysis==1.0.0b1
pytest==7.4.0
black==23.9.1
flake8==6.1.0
pytest-cov==4.1.0
types-pillow==10.2.0.20240423
mypy==1.8.0